Leasey-Team13-main 3/backend/requirements.txt
boto3==1.34.80
botocore==1.34.80
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==2.2.5
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
pandas==2.0.3
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
setuptools==69.2.0
six==1.16.0
urllib3== 1.26.18
Werkzeug==2.2.2
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
PyTest==8.1.1
moto==5.0.5
cryptography==41.0.7
